Implemented --watch switch
gsassc watches for changes either in a single file or
in a directory. The implementation is as close as i can
tell it is in the original (ruby) sass.

If gsassc is watching a directory and the changed file is
a partial (filename starting with a underscore) it will
search for all non-partials (no underscroe, .scss suffix) in
the watched dir and compile the first one.